modified:   package-lock.json 	modified:   package.json 	deleted:    src/App.js 	new file:   src/App.jsx 	renamed:    src/components/Home.js -> src/components/Home.jsx 	renamed:    src/components/SidePanel.js -> src/components/SidePanel.jsx 	renamed:    src/components/Skills.js -> src/components/Skills.jsx 	renamed:    src/components/Thoughts.js -> src/components/Thoughts.jsx 	renamed:    src/index.js -> src/index.jsx 	modified:   src/styles/home.css 	modified:   src/styles/thoughts.css
