+++ src/components/Home.jsx
@@ -2,6 +2,7 @@ import React from "react";
 import "../styles/home.css";
 import image from "../images/image.png";
 import { Thoughts } from "./Thoughts";
+import { Link } from "react-router-dom";
 export const Home = () => {
   return (
     <div className="home">
@@ -13,22 +14,23 @@ export const Home = () => {
         <h3>Hello there !</h3>
         <h4>I'm Vigneshwaran</h4>
         <p>
-          Hey there! I'm your friendly neighborhood full-stack JavaScript
-          junkie, always hungry for new challenges! Whether I'm crushing code as
-          a full-stack developer, soaking up knowledge as a student, or crafting
-          conversational magic with chatbots, one thing's for sure – I live,
-          breathe, hate and love JS
+          Hey there! I'm your <span>friendly neighborhood</span> full-stack{" "}
+          <span>JavaScript </span> junkie, always hungry for new challenges!
+          Whether I'm crushing code as a <span>full-stack developer</span>,
+          soaking up knowledge as a <span>student</span>, or crafting
+          conversational magic with <span>chatbots</span>, one thing's for sure
+          – I live, breathe, <span>hate and love JS</span>
         </p>
       </div>
       <div className="butt">
-        <a
+        <Link
           class="inline-flex items-center px-4 py-2 mt-2 font-semibold tracking-tighter text-white transition duration-500 ease-in-out transform bg-transparent border rounded-lg text-md md:mt-0 hover:text-black hover:bg-white focus:shadow-outline"
-          href="/"
+          to={"/explore"}
         >
           <div class="flex text-lg tracking-tighter">
             <span class="justify-center">Explore </span>
           </div>
-        </a>
+        </Link>
       </div>
       <Thoughts />
       <div></div>
