+++ src/components/Thoughts.jsx
@@ -5,8 +5,11 @@ export const Thoughts = () => {
     <div className="thoughts">
       <h1>Me currently : </h1>
       <ul>
-        <li>"Life is getting Fucked up fellas"</li>
-        <li>"Life is getting Fucked up fellas"</li>
+        <li>" Started having interest in React Native, Gonna try it out "</li>
+        <li>
+          me the entire dev process : "I Hate JS" <br /> me after deployment :
+          "Oh i love JS"
+        </li>
         <li>"Life is getting Fucked up fellas"</li>
       </ul>
     </div>
