+++ src/styles/home.css
@@ -1,5 +1,5 @@
 .home {
-  height: 100%;
+  height: 100vh;
   width: 50%;
 }
 .butt {
@@ -63,4 +63,19 @@
   margin-left: 10%;
   font-size: 13px;
   color: rgb(99, 99, 99);
+}
+.intro p span {
+  color: rgb(199, 199, 199);
+}
+
+@media only screen and (max-width: 999px) {
+  .home {
+    overflow-y: hidden;
+    .logo {
+      position: absolute;
+    }
+    .intro {
+      min-width: 320px;
+    }
+  }
 }
