+++ src/styles/thoughts.css
@@ -26,4 +26,5 @@
   margin-left: 10%;
   border-radius: 20px;
   margin: 20px auto;
+  padding: 0px 20px;
 }
